// src/modules/attendance/dto/finalized-payable-days.dto.ts
export class FinalizedPayableDaysDto {
  employeeId: string;
  employeeCode: string;
  employeeName: string;
  year: number;
  month: number;
  totalMonthDays: number;
  presentDays: number;
  weeklyOffDays: number;
  paidLeaveDays: number;
  halfDays: number;
  unexcusedAbsenceDays: number;
  lwpDays: number;
  payableDays: number;
}